mengubah penempatan code & menambahkan deskripsi program
mengubah penempatan code agar lebih tertata rapih dan menambahkan beberapa keterangan pada program.
tolong tambahkan deskripsi per modul nya

Note tambahan : masih terdapat bug, saat user memasukkan input yg tidak diinginkan program akan crash
